BJS2-96643 get users controllers review fixes
